LibFT: ft_numlen takes uint64_t

+++ libft/ft_numlen.c
@@ -6,12 +6,12 @@
 /*   By: chamada <chamada@student.le-101.fr>        +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/11/22 23:14:04 by chamada      #+#   ##    ##    #+#       */
-/*   Updated: 2019/11/22 23:24:56 by chamada     ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/11/23 18:55:28 by chamada     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
-int	ft_numlen(long n, unsigned base)
+int	ft_numlen(unsigned long long n, unsigned base)
 {
 	int	len;
